+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## v3.2.0
 
+### `@liveblocks/react-ui`
+
+- Add `blurOnSubmit` prop to `Composer` (also available on the `Composer.Form`
+  primitive and as `blurComposerOnSubmit` on `Thread`) to control whether a
+  composer should lose focus after being submitted.
+
 ## v3.1.3
 
 ### `@liveblocks/react-ui`